Detach the OutputView as Admin and User #23
OutputView를 사용자와 관리자로 분리해서 출력하도록 수정했다.
